Move Country and Language into indigo_api
See https://stackoverflow.com/questions/30601107/move-models-between-django-1-8-apps-with-required-foreignkey-references

Ref #148
