+++ indigo_app/context_processors.py
@@ -19,7 +19,7 @@ def general(request):
 def models(request):
     """ Add some useful models to templates
     """
-    from indigo_app.models import Country, Language
+    from indigo_api.models import Country, Language
 
     return {
         'indigo_languages': Language.objects,
