+++ indigo_app/models.py
@@ -3,65 +3,15 @@ from django.contrib.auth.models import User
 from django.db.models.signals import post_save, pre_save
 from django.dispatch import receiver
 
-from languages_plus.models import Language as MasterLanguage
-from countries_plus.models import Country as MasterCountry
 from rest_framework.authtoken.models import Token
 
-from indigo_api.models import Document
-
-
-class Language(models.Model):
-    """ The languages available in the UI. They aren't enforced by the API.
-    """
-    language = models.OneToOneField(MasterLanguage, on_delete=models.CASCADE)
-
-    class Meta:
-        ordering = ['language__name_en']
-
-    def __unicode__(self):
-        return unicode(self.language)
-
-
-class Country(models.Model):
-    """ The countries available in the UI. They aren't enforced by the API.
-    """
-    country = models.OneToOneField(MasterCountry, on_delete=models.CASCADE)
-    primary_language = models.ForeignKey(Language, on_delete=models.PROTECT, null=False, related_name='+', help_text='Primary language for this country')
-
-    class Meta:
-        ordering = ['country__name']
-        verbose_name_plural = 'Countries'
-
-    @property
-    def code(self):
-        return self.country.iso.lower()
-
-    @property
-    def name(self):
-        return self.country.name
-
-    def as_json(self):
-        return {
-            'name': self.name,
-            'localities': {loc.code: loc.name for loc in self.locality_set.all()},
-            'publications': [pub.name for pub in self.publication_set.all()],
-        }
-
-    def work_locality(self, work):
-        return self.locality_set.filter(code=work.locality).first()
-
-    def __unicode__(self):
-        return unicode(self.country.name)
-
-    @classmethod
-    def for_work(cls, work):
-        return cls.objects.select_related('country').filter(country__iso__iexact=work.country).first()
+from indigo_api.models import Document, Country
 
 
 class Locality(models.Model):
     """ The localities available in the UI. They aren't enforced by the API.
     """
-    country = models.ForeignKey(Country, null=False, on_delete=models.CASCADE)
+    country = models.ForeignKey('indigo_api.Country', null=False, on_delete=models.CASCADE)
     name = models.CharField(max_length=512, null=False, blank=False, help_text="Local name of this locality")
     code = models.CharField(max_length=100, null=False, blank=False, help_text="Unique code of this locality (used in the FRBR URI)")
 
@@ -85,7 +35,7 @@ class Editor(models.Model):
     properties that we need, like a default country.
     """
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    country = models.ForeignKey(Country, on_delete=models.SET_NULL, null=True)
+    country = models.ForeignKey('indigo_api.Country', on_delete=models.SET_NULL, null=True)
     accepted_terms = models.BooleanField(default=False)
 
     @property
@@ -109,7 +59,7 @@ class Editor(models.Model):
 class Publication(models.Model):
     """ The publications available in the UI. They aren't enforced by the API.
     """
-    country = models.ForeignKey(Country, null=False, on_delete=models.CASCADE)
+    country = models.ForeignKey('indigo_api.Country', null=False, on_delete=models.CASCADE)
     name = models.CharField(max_length=512, null=False, blank=False, unique=True, help_text="Name of this publication")
 
     class Meta:
